refactor: moved data tab components to pages/map/data folder
