+++ sites/geohub/src/components/Content.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { page } from '$app/stores';
-	import DataView from '$components/DataView.svelte';
+	import DataView from '$components/pages/map/data/DataView.svelte';
 	import LayerList from '$components/LayerList.svelte';
 	import { TabNames } from '$lib/config/AppConfig';
 	import { handleEnterKey } from '$lib/helper';
