+++ sites/geohub/src/components/pages/map/data/DataCardInfo.svelte
