+++ sites/geohub/src/components/pages/map/data/DataCategoryCard.svelte
