+++ sites/geohub/src/components/util/LayerTypeSwitch.svelte
